health-checker/requirements.txt
requests
pyyaml
slackclient==2.9.4